+++ app/src/components/Header.jsx
@@ -4,9 +4,12 @@ import '../css/Header.css'
 const Header = () => {
   return (
     <div className='Header'>
-      <div className='Title'>Birdsnest</div>
-      <div className='MenuBtn'>Menu</div>
-      <div></div>
+      <h1 className='Title'>
+        Birdsnest
+      </h1>
+      <h2 className='MenuBtn'>
+        Menu
+      </h2>
     </div>
   )
 }
