+++ app/src/css/App.css
@@ -12,6 +12,8 @@
 
 .App {
   text-align: center;
+  min-height: 100vh;
+  background-color: var(--dark);
 }
 
 .App-logo {
